Started Implementing Question Datastructures
Questions are now handled using an Interface that defines the need for a prompt and a number of points, and then by their own classes, one for each type, each one holding it's own specific data. In this case the MultChoiceQuestion class contains all the information for a multiple choice question; The required prompt and point value from the interface along with an array of answers.